+++ lib/src/chain.dart
@@ -72,20 +72,23 @@ class Chain implements StackTrace {
   ///  If [errorZone] is `false`, [onError] must be `null`.
   ///
   /// If [callback] returns a value, it will be returned by [capture] as well.
+  ///
+  /// [zoneValues] is added to the [runZoned] calls.
   static T capture<T>(T Function() callback,
       {void Function(Object error, Chain)? onError,
       bool when = true,
-      bool errorZone = true}) {
+      bool errorZone = true,
+      Map<Object?, Object?>? zoneValues}) {
     if (!errorZone && onError != null) {
       throw ArgumentError.value(
           onError, 'onError', 'must be null if errorZone is false');
     }
 
     if (!when) {
-      if (onError == null) return runZoned(callback);
+      if (onError == null) return runZoned(callback, zoneValues: zoneValues);
       return runZonedGuarded(callback, (error, stackTrace) {
         onError(error, Chain.forTrace(stackTrace));
-      }) as T;
+      }, zoneValues: zoneValues) as T;
     }
 
     var spec = StackZoneSpecification(onError, errorZone: errorZone);
@@ -102,9 +105,11 @@ class Chain implements StackTrace {
         // where T is a nullable type continue to work.
         return null as T;
       }
-    },
-        zoneSpecification: spec.toSpec(),
-        zoneValues: {_specKey: spec, StackZoneSpecification.disableKey: false});
+    }, zoneSpecification: spec.toSpec(), zoneValues: {
+      ...?zoneValues,
+      _specKey: spec,
+      StackZoneSpecification.disableKey: false
+    });
   }
 
   /// If [when] is `true` and this is called within a [Chain.capture] zone, runs